+++ tests/test_clvm_serde.py
@@ -7,18 +7,18 @@ from chia_base.core import Coin, CoinSpend
 
 from clvm_rs import Program
 
+from hsms.clvm_serde import (
+    to_program_for_type,
+    from_program_for_type,
+    Frugal,
+    tuple_frugal,
+)
 from hsms.core.signing_hints import SumHint, PathHint
 from hsms.core.unsigned_spend import (
     UnsignedSpend,
     from_storage,
     to_storage,
     TO_PROGRAM,
-)
-from hsms.util.clvm_serde import (
-    to_program_for_type,
-    from_program_for_type,
-    Frugal,
-    tuple_frugal,
 )
 from .legacy.signing_hints import (
     SumHint as LegacySH,
